Add more hotkeys + update readme
